+++ media/webrtc/trunk/webrtc/modules/audio_coding/codecs/g722/g722_enc_dec.h
@@ -31,7 +31,7 @@
 #if !defined(_G722_ENC_DEC_H_)
 #define _G722_ENC_DEC_H_
 
-#include "webrtc/typedefs.h"
+#include "typedefs.h"  // NOLINT(build/include)
 
 /*! \page g722_page G.722 encoding and decoding
 \section g722_page_sec_1 What does it do?
@@ -139,19 +139,19 @@ G722EncoderState* WebRtc_g722_encode_init(G722EncoderState* s,
                                           int rate,
                                           int options);
 int WebRtc_g722_encode_release(G722EncoderState *s);
-int WebRtc_g722_encode(G722EncoderState *s,
-                       uint8_t g722_data[],
-                       const int16_t amp[],
-                       int len);
+size_t WebRtc_g722_encode(G722EncoderState *s,
+                          uint8_t g722_data[],
+                          const int16_t amp[],
+                          size_t len);
 
 G722DecoderState* WebRtc_g722_decode_init(G722DecoderState* s,
                                           int rate,
                                           int options);
 int WebRtc_g722_decode_release(G722DecoderState *s);
-int WebRtc_g722_decode(G722DecoderState *s,
-                       int16_t amp[],
-                       const uint8_t g722_data[],
-                       int len);
+size_t WebRtc_g722_decode(G722DecoderState *s,
+                          int16_t amp[],
+                          const uint8_t g722_data[],
+                          size_t len);
 
 #ifdef __cplusplus
 }
