+++ media/webrtc/trunk/webrtc/modules/audio_coding/codecs/isac/fix/source/lattice_mips.c
@@ -8,8 +8,10 @@
  *  be found in the AUTHORS file in the root of the source tree.
  */
 
-#include "webrtc/modules/audio_coding/codecs/isac/fix/source/settings.h"
-#include "webrtc/typedefs.h"
+#include <stddef.h>
+
+#include "modules/audio_coding/codecs/isac/fix/source/settings.h"
+#include "typedefs.h"  // NOLINT(build/include)
 
 // Filter ar_g_Q0[] and ar_f_Q0[] through an AR filter with coefficients
 // cth_Q15[] and sth_Q15[].
@@ -17,11 +19,11 @@ void WebRtcIsacfix_FilterArLoop(int16_t* ar_g_Q0,     // Input samples
                                 int16_t* ar_f_Q0,     // Input samples
                                 int16_t* cth_Q15,     // Filter coefficients
                                 int16_t* sth_Q15,     // Filter coefficients
-                                int16_t order_coef) { // order of the filter
+                                size_t order_coef) { // order of the filter
   int n = 0;
 
   for (n = 0; n < HALF_SUBFRAMELEN - 1; n++) {
-    int count = order_coef - 1;
+    int count = (int)(order_coef - 1);
     int offset;
 #if !defined(MIPS_DSP_R1_LE)
     int16_t* tmp_cth;
